+++ src/renderer/components/icons/index.ts
@@ -19,6 +19,7 @@ import userAvatarIcon from '../../assets/svg/asset-user.svg'
 import xRuneIcon from '../../assets/svg/asset-xrune.svg'
 import { ReactComponent as CurrencyIcon } from '../../assets/svg/currency-icon.svg'
 import blueArrowIcon from '../../assets/svg/icon-arrow-button.svg'
+import { ReactComponent as AttentionIcon } from '../../assets/svg/icon-attention.svg'
 import { ReactComponent as LedgerIcon } from '../../assets/svg/ledger.svg'
 import orbBlueIcon from '../../assets/svg/orb-blue.svg'
 import orbGreenIcon from '../../assets/svg/orb-green.svg'
@@ -46,5 +47,6 @@ export {
   blackArrowIcon,
   blueArrowIcon,
   greyArrowIcon,
-  LedgerIcon
+  LedgerIcon,
+  AttentionIcon
 }