+++ src/renderer/i18n/en/common.ts
@@ -94,7 +94,8 @@ const common: CommonMessages = {
   'common.informationMore': 'More information',
   'common.balance': 'Balance',
   'common.balance.loading': 'Loading balance',
-  'common.custom': 'Custom'
+  'common.custom': 'Custom',
+  'common.notsupported.fornetwork': 'Not supported for {network}'
 }
 
 export default common